backend/pyproject.toml
[project]
name = "storystitch-backend"
version = "0.1.0"
description = "StoryStitch FastAPI backend — OpenAI (text), Gemini (images), ElevenLabs (voice)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "openai>=1.30.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
